# perf: Process chunks of bracketed pastes a linear number of times.