--- a/src/Graphics/Vty/Input/Classify/Types.hs
+++ b/src/Graphics/Vty/Input/Classify/Types.hs
@@ -8,8 +8,10 @@ where
 
 import Graphics.Vty.Input.Events
 
+import Data.ByteString.Char8 (ByteString)
+
 data KClass
-    = Valid Event String
+    = Valid Event ByteString
     -- ^ A valid event was parsed. Any unused characters from the input
     -- stream are also provided.
     | Invalid