@iphydf 2022-03-10
I finished my own review. This is ready for review.

I was a bit hesitant on renaming `dropInvalid`, because it does more now than just dropping invalid sequences (it moves chunks from unprocessed to classifier state), but I kept the name. If you have a better proposal, I can rename it.