@iphydf 2022-03-10
In fact, my timings are likely very far off: my Haskell process is only using 6% CPU for the duration of that paste. Now it's network and terminal throughput bound.