@jtdaugherty 2022-03-10
This looks okay to me. I also tested it in moderate anger with [matterhorn](https://github.com/matterhorn-chat/matterhorn) and didn't run into any issues. As for `dropInvalid`, I would like to come up with one that better describes what it is taking care of, but I don't have any ideas right now. I'm going to merge this. Thanks again for your investigations, detailed findings, and patches!